Implement packet reorder buffer

+++ src/reorder.c
@@ -5,7 +5,7 @@ void reorder_init()
 }
 
 int reorder_packet(
-    unsigned short packet_id,
+    unsigned int packet_id,
     unsigned char *packet,
     int length,
     unsigned char **out_packet,
